+++ example/lib/super_auto_suggestions_item_scenarios_screen.dart
@@ -58,6 +58,7 @@ class _SuperAutoSuggestionsItemScenariosScreenState
   }
 
   SuperAutoSuggestionsItem<_ScenarioItem> _suggestion(
+    BuildContext context,
     List<_ScenarioItem> items,
     int index,
     _ScenarioItem item,